requirements.txt
astroid==2.9.3
autopep8==1.6.0
exif==1.3.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
piexif==1.1.3
Pillow==9.0.1
platformdirs==2.5.0
plum-py==0.7.3
pycodestyle==2.8.0
pylint==2.12.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
toml==0.10.2
typing_extensions==4.1.0
wrapt==1.13.3
